+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 ### Added
 - Ability to get the selected backend by reading `ngl.Config.backend` after the
   configure step, in Python
+- `ngl.SceneCfg` now includes the backend capabilities
 
 ### Fixed
 - Moving the split position in `ngl-diff`